修改tappay key存放位置、會員登入註冊前端驗證、修改選擇日期視窗(不能選擇過去日期)
