+++ static/css/attraction.css
@@ -71,19 +71,7 @@
   position: relative;
 }
 
-.username {
-  padding-left: 15px;
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 13px;
-}
-.email {
-  padding-left: 15px;
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 13px;
-}
-.password {
+.sign1 {
   padding-left: 15px;
   font-weight: 400;
   font-size: 16px;
@@ -154,7 +142,7 @@ input[type="date"]::-webkit-calendar-picker-indicator {
   color: rgba(0, 0, 0, 0);
   opacity: 1;
   display: block;
-  background: url("../static/pic/icon_calendar.png") no-repeat;
+  background: url("../pic/icon_calendar.png") no-repeat;
   width: 20px;
   height: 20px;
   border-width: thin;
@@ -502,7 +490,7 @@ input[type="date"]::-webkit-calendar-picker-indicator {
   #ball {
     position: absolute;
     top: 91%;
-    left: 29%;
+    left: 26%;
   }
   .upper {
     display: block;
@@ -550,6 +538,11 @@ input[type="date"]::-webkit-calendar-picker-indicator {
   }
   .line {
     width: 400px;
+  }
+}
+@media screen and (max-width: 520px){
+  #ball {
+    left: 21%;
   }
 }
 @media screen and (max-width: 450px) {
