+++ static/css/booking.css
@@ -289,7 +289,6 @@ input {
   }
   .first_part {
     width: 330px;
-    height: 1065px;
   }
   .line {
     width: 330px;
