+++ static/js/index.js
@@ -3,7 +3,10 @@ let data=null;
 let loading=false;
 let statusData;
 
-//載入首頁的資料
+//先載入首頁資料
+getData(0);
+
+//載入首頁資料函式
 function getData(page){
     //檢查會員登入狀態
     let preSignin=document.getElementById("preSignin");
@@ -25,7 +28,6 @@ function getData(page){
             bookreq.onload=function(){
             prebookData=JSON.parse(bookreq.responseText);
             let count=prebookData.count;
-            console.log(prebookData)
             if(count != null){
                 let img=document.createElement("img");
                 img.setAttribute("class","bookcarticon");
@@ -43,7 +45,7 @@ function getData(page){
     }
     
     loading=true;
-    fetch("http://18.180.51.21:3000/api/attractions?page="+page+"&keyword=")
+    fetch("/api/attractions?page="+page+"&keyword=")
     .then(function(response){
         return response.json();
     }).then(function(result){
@@ -101,18 +103,20 @@ function getData(page){
         document.querySelector("#loading").style.display="none";
 }
 )
-}
-
-//首頁資料的scroll事件  
-if (data == null){
-document.addEventListener("scroll", lazyLoad)
-function lazyLoad() {
-    if (window.pageYOffset + window.innerHeight >= document.documentElement.scrollHeight && loading==false) {                
-        if (next_page==null){document.removeEventListener("scroll", lazyLoad)}
-        else{getData(next_page);console.log(next_page)};
-        }  
+    //首頁資料的scroll事件  
+    if(data == null){
+    document.addEventListener("scroll", lazyLoad)
+    function lazyLoad() {
+        if (window.pageYOffset + window.innerHeight >= document.documentElement.scrollHeight && loading==false) { 
+            console.log(data)
+            if (next_page==null){document.removeEventListener("scroll", lazyLoad)}
+            else{getData(next_page)};
+            }  
+    }
     }
 }
+
+
 
 //搜尋關鍵字後先把首頁資料移除
 function remove(){
@@ -128,13 +132,14 @@ function query(anotherpage){
     next_page==null;
     data=document.getElementById("query").value;
     loading=true;
-    fetch("http://18.180.51.21:3000/api/attractions?page="+anotherpage+"&keyword="+data)
+    fetch("/api/attractions?page="+anotherpage+"&keyword="+data)
     .then(function(response){
         return response.json();
     }).then(function(result){
         if(result.error==true){document.getElementById("attractions").innerHTML="查無景點資料";}
         let url;
         queryNextpage=result.nextPage;
+        console.log(queryNextpage)
         let fragment=document.createDocumentFragment();
         function length(obj) {
             return Object.keys(obj).length;
@@ -183,16 +188,17 @@ function query(anotherpage){
         loading=false;
 }
 )
-}
-
-//關鍵字資料的scroll事件
-document.addEventListener("scroll", lazyLoad2)
-function lazyLoad2() {
-    if (window.pageYOffset + window.innerHeight >= document.documentElement.scrollHeight && loading==false) {                
-        if (queryNextpage==null){document.removeEventListener("scroll", lazyLoad2)}
-        else{query(queryNextpage);console.log(queryNextpage)};
+    //關鍵字資料的scroll事件
+    document.addEventListener("scroll", lazyLoad2)
+    function lazyLoad2() {
+        if (window.pageYOffset + window.innerHeight >= document.documentElement.scrollHeight && loading==false) {    
+            console.log(data)
+            if (queryNextpage==null){document.removeEventListener("scroll", lazyLoad2)}
+            else{query(queryNextpage)};
         }  
     }
+}
+
 
 
 //點選導覽列的 預定行程 處理
