+++ templates/base.html
@@ -6,9 +6,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="icon" type="image/png" sizes="32x32" href="../static/pic/bx-building-house.svg">
     <title>台北一日遊</title>
-    <link href="../static/css/index.css" rel="stylesheet"/> 
     <link href="../static/css/same.css" rel="stylesheet"/>
-    <link href="../static/css/signinup.css" rel="stylesheet"/>
+    {% block head %}{% endblock %}
 </head>
 <body onselectstart="return false">
     <div id="signin">
@@ -65,24 +64,10 @@
             <a class="righttitle2" onclick="tosignin_out()" id="afterSignin">登出系統</a>
         </nav>
     </div>
-   
-    <div class="welcome">
-        <div class="subtitle">輕鬆享受台北一日悠閒</div>
-        <div class="slogan">探索每個角落，體驗城市的深度旅遊行程</div>
-        <div class="query">
-        <input type="text" placeholder="輸入景點名稱查詢" id="query">
-        <button class="button" onclick="remove(query(0))"><img src="../static/pic/icon_search.png" class="img1" /></button>
-        </div>
-        
-        <img src="../static/pic/welcome.png" class="picture"> 
-    </div>
-    <div id="attractions" class="attractions">
-    </div>
+    {% block content %}{% endblock %}
     <div class="footer">
         <div class="copyright" id="copyright">COPYRIGHT © 2021 台北一日遊</div>
     </div>
-    
-    <script type="text/javascript" src="../static/js/index.js"></script>
-    <script type="text/javascript" src="../static/js/LoginoutRegister.js"></script>
+    {% block js %}{% endblock %}
 </body>
 </html>
