+++ templates/booking.html
@@ -6,8 +6,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="icon" type="image/png" sizes="32x32" href="../static/pic/bx-building-house.svg">
     <title>台北一日遊</title>
-    <link href="../static/booking.css" rel="stylesheet"/>
-    <link href="../static/same.css" rel="stylesheet"/>
+    <link href="../static/css/booking.css" rel="stylesheet"/>
+    <link href="../static/css/same.css" rel="stylesheet"/>
     <script src="https://js.tappaysdk.com/tpdirect/v5.8.0"></script>
 </head>
 <body onload="checkStatus()" ; onselectstart="return false">
@@ -26,19 +26,19 @@
         <div id="bookingdata"></div>
     </div>
     <div class="line" id="line1"></div>
+    <form onsubmit="gopay()">
     <div class="second_part" id="second">
         <div class="contact">
         <div class="contactdata">您的聯絡資訊</div>
-        <div class="contactname">聯絡姓名：<input type="text" id="contactperson"></div>
-        <div class="contactemail">連絡信箱：<input type="text" id="contact_email"></div>
-        <div class="contactnumber">手機號碼：<input type="tel" required pattern="^09[0-9]{8}$" maxlength="10" id="contact_tel"></div>
+        <div class="contactname">聯絡姓名：<input type="text" id="contactperson" required></div>
+        <div class="contactemail">連絡信箱：<input type="email" id="contact_email" required></div>
+        <div class="contactnumber">手機號碼：<input type="tel" required pattern="^09[0-9]{8}$" maxlength="10" id="contact_tel" required></div>
         <div class="notice">請保持手機暢通，準時到達，導覽人員將用手機與您聯繫，務必留下正確的聯絡方式。</div>
         </div>
     </div>
     <div class="line" id="line2"></div>
-    <form onsubmit="gopay()">
+    
         <div class="third_part" id="third">
-            
             <div class="pay">
             <div class="creditcard">信用卡付款資訊</div>
             <div class="form-group card-number-group">
@@ -66,7 +66,7 @@
     <div class="footer" id="footer">
         <div class="copyright" id="copyright">COPYRIGHT © 2021 台北一日遊</div>
     </div>
-    <script type="text/javascript" src="../static/booking.js"></script>
-    <script type="text/javascript" src="../static/validate.js"></script>
+    <script type="text/javascript" src="../static/js/booking.js"></script>
+    <script type="text/javascript" src="../static/js/validate.js"></script>
 </body>
 </html>
