+++ templates/thankyou.html
@@ -6,8 +6,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="icon" type="image/png" sizes="32x32" href="../static/pic/bx-building-house.svg">
     <title>台北一日遊</title>
-    <link href="../static/same.css" rel="stylesheet"/>
-    <link href="../static/thankyou.css" rel="stylesheet"/>
+    <link href="../static/css/same.css" rel="stylesheet"/>
+    <link href="../static/css/thankyou.css" rel="stylesheet"/>
 </head>
 <body onload="checkStatus()" ; onselectstart="return false">
     <div id="loading"><img src="../static/pic/loading1.gif" id="loadingicon"></div>
